LPD-25361 - Rename
