+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/action/trigger/messaging/ObjectActionTriggerMessageListener.java
@@ -6,55 +6,200 @@
 package com.liferay.object.internal.action.trigger.messaging;
 
 import com.liferay.object.action.engine.ObjectActionEngine;
-import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.object.constants.ObjectActionTriggerConstants;
+import com.liferay.object.internal.entry.util.ObjectEntryUtil;
+import com.liferay.object.model.ObjectDefinition;
+import com.liferay.object.model.ObjectEntry;
+import com.liferay.object.service.ObjectDefinitionLocalService;
+import com.liferay.object.service.ObjectEntryLocalService;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
+import com.liferay.portal.kernel.json.JSONFactory;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.messaging.BaseMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationConfiguration;
+import com.liferay.portal.kernel.messaging.DestinationFactory;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.messaging.MessageListener;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.MapUtil;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.vulcan.dto.converter.DTOConverterRegistry;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceRegistration;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Marco Leo
- * @author Brian Wing Shun Chan
+ * @author Marco Leov
  */
+@Component(
+	property = {
+		"destination.name=" + DestinationNames.OBJECT_ENTRY_ATTACHMENT_DOWNLOAD,
+		"destination.name=" + DestinationNames.OBJECT_ENTRY_ON_AFTER_ADD,
+		"destination.name=" + DestinationNames.OBJECT_ENTRY_ON_AFTER_UPDATE
+	},
+	service = MessageListener.class
+)
 public class ObjectActionTriggerMessageListener extends BaseMessageListener {
 
-	public ObjectActionTriggerMessageListener(
-		String className, ObjectActionEngine objectActionEngine,
-		String objectActionTriggerKey) {
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_registerService(
+			bundleContext, DestinationNames.OBJECT_ENTRY_ON_AFTER_ADD);
 
-		_className = className;
-		_objectActionEngine = objectActionEngine;
-		_objectActionTriggerKey = objectActionTriggerKey;
+		_registerService(
+			bundleContext, DestinationNames.OBJECT_ENTRY_ON_AFTER_UPDATE);
+
+		_registerService(
+			bundleContext, DestinationNames.OBJECT_ENTRY_ATTACHMENT_DOWNLOAD);
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceRegistrations.forEach(ServiceRegistration::unregister);
+
+		_serviceRegistrations.clear();
 	}
 
 	@Override
-	protected void doReceive(Message message) {
+	protected void doReceive(Message message) throws Exception {
+		String objectDefinitionExternalReferenceCode = message.getString(
+			"objectDefinitionExternalReferenceCode");
+		long companyId = message.getLong("companyId");
+
+		ObjectDefinition objectDefinition =
+			_objectDefinitionLocalService.
+				fetchObjectDefinitionByExternalReferenceCode(
+					objectDefinitionExternalReferenceCode, companyId);
+
+		if (objectDefinition == null) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(
+					StringBundler.concat(
+						"Object definition is null for external reference ",
+						"code ", objectDefinitionExternalReferenceCode,
+						" and company ", companyId));
+
+				return;
+			}
+		}
+
+		String objectEntryExternalReferenceCode = message.getString(
+			"objectEntryExternalReferenceCode");
+
+		ObjectEntry objectEntry = _objectEntryLocalService.fetchObjectEntry(
+			objectEntryExternalReferenceCode,
+			objectDefinition.getObjectDefinitionId());
+
+		if (objectEntry == null) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(
+					StringBundler.concat(
+						"Object entry is null for external reference code ",
+						objectEntryExternalReferenceCode,
+						" and object definition ",
+						objectDefinitionExternalReferenceCode));
+
+				return;
+			}
+		}
+
+		String objectActionTrigger = message.getString("objectActionTrigger");
+
 		_objectActionEngine.executeObjectActions(
-			_className, GetterUtil.getLong(message.get("companyId")),
-			_objectActionTriggerKey, () -> (JSONObject)message.getPayload(),
-			_getUserId(message));
-	}
+			objectDefinition.getClassName(), message.getLong("companyId"),
+			objectActionTrigger,
+			() -> ObjectEntryUtil.getPayloadJSONObject(
+				_dtoConverterRegistry, _jsonFactory, objectActionTrigger,
+				objectDefinition, objectEntry, null,
+				_userLocalService.getUser(message.getLong("userId"))),
+			message.getLong("userId"));
 
-	private long _getUserId(Message message) {
-		long userId = GetterUtil.getLong(message.get("principalName"));
+		if (!FeatureFlagManagerUtil.isEnabled("LPS-187142") ||
+			(!objectDefinition.isRootDescendantNode() &&
+			 (!objectDefinition.isRootNode() ||
+			  StringUtil.equals(
+				  objectActionTrigger,
+				  ObjectActionTriggerConstants.KEY_ON_AFTER_ADD)))) {
 
-		if (userId != 0L) {
-			return userId;
+			return;
 		}
 
-		Object object = message.get("permissionChecker");
+		ObjectEntry rootObjectEntry = _objectEntryLocalService.fetchObjectEntry(
+			objectEntry.getRootObjectEntryId());
 
-		if (!(object instanceof PermissionChecker)) {
-			return 0L;
+		if (rootObjectEntry == null) {
+			return;
 		}
 
-		PermissionChecker permissionChecker = (PermissionChecker)object;
+		ObjectDefinition rootObjectDefinition =
+			_objectDefinitionLocalService.getObjectDefinition(
+				rootObjectEntry.getObjectDefinitionId());
 
-		return permissionChecker.getUserId();
+		_objectActionEngine.executeObjectActions(
+			rootObjectDefinition.getClassName(), message.getLong("companyId"),
+			ObjectActionTriggerConstants.KEY_ON_AFTER_ROOT_UPDATE,
+			() -> ObjectEntryUtil.getPayloadJSONObject(
+				_dtoConverterRegistry, _jsonFactory,
+				ObjectActionTriggerConstants.KEY_ON_AFTER_ROOT_UPDATE,
+				rootObjectDefinition, rootObjectEntry, null,
+				_userLocalService.getUser(message.getLong("userId"))),
+			message.getLong("userId"));
 	}
 
-	private final String _className;
-	private final ObjectActionEngine _objectActionEngine;
-	private final String _objectActionTriggerKey;
+	private void _registerService(
+		BundleContext bundleContext, String destinationName) {
+
+		DestinationConfiguration destinationConfiguration =
+			new DestinationConfiguration(
+				DestinationConfiguration.DESTINATION_TYPE_SYNCHRONOUS,
+				destinationName);
+
+		Destination destination = _destinationFactory.createDestination(
+			destinationConfiguration);
+
+		_serviceRegistrations.add(
+			bundleContext.registerService(
+				Destination.class, destination,
+				MapUtil.singletonDictionary(
+					"destination.name", destination.getName())));
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		ObjectActionTriggerMessageListener.class);
+
+	@Reference
+	private DestinationFactory _destinationFactory;
+
+	@Reference
+	private DTOConverterRegistry _dtoConverterRegistry;
+
+	@Reference
+	private JSONFactory _jsonFactory;
+
+	@Reference
+	private ObjectActionEngine _objectActionEngine;
+
+	@Reference
+	private ObjectDefinitionLocalService _objectDefinitionLocalService;
+
+	@Reference
+	private ObjectEntryLocalService _objectEntryLocalService;
+
+	private final List<ServiceRegistration<Destination>> _serviceRegistrations =
+		new ArrayList<>();
+
+	@Reference
+	private UserLocalService _userLocalService;
 
 }
